Add async job processing
